Everything works. Tryingsolution on orginal layout
