+++ src/components/Week.vue
@@ -2,63 +2,38 @@
   <div :key="day.id" v-for="day in week">
     <Day
       @toggle-complete-day="$emit('toggle-complete-day', day.id)"
-      @toggle-complete="toggleComplete"
       :day="day"
-      :lifts="lifts"
     />
+    <div :key="lift.id" v-for="lift in JSON.parse(JSON.stringify(lifts))">
+      <div
+        v-if="
+          day.id === JSON.parse(JSON.stringify(lifts[lift.id - 1].parentId))
+        "
+      >
+        <Lift
+          @toggle-complete="$emit('toggle-complete', lift.id)"
+          :lift="lift"
+        />
+      </div>
+    </div>
   </div>
 </template>
 
 <script>
 import { Vue } from "vue";
 import Day from "./Day";
+import Lift from "./Lift";
 export default {
   name: "Week",
   props: {
     week: Array,
     lifts: Array,
+    lift: Object,
   },
   components: {
     Day,
-  },
-  data() {
-    return {
-      bla: [],
-      lift: Object,
-    };
+    Lift,
   },
   emits: ["toggle-complete", "toggle-complete-day", "toggleComplete"],
-  methods: {
-    async toggleComplete(id) {
-      const LiftToToggle = await this.fetchLift(id);
-      const updLift = { ...LiftToToggle, complete: !LiftToToggle.complete };
-      var res = await fetch(`api/lifts/${id}`, {
-        method: "PUT",
-        headers: {
-          "Content-type": "application/json",
-        },
-        body: JSON.stringify(updLift),
-      });
-
-      const data = await res.json();
-      this.lift = Object.assign({}, this.lift, {
-        complete: data.complete,
-      });
-      console.log(this.lift);
-      /*bla = this.lifts.map((lift) =>
-        lift.id === id ? { ...lift, complete: data.complete } : lift
-      );*/
-    },
-    async created() {
-      this.lifts = await this.fetchLifts();
-    },
-    async fetchLift(id) {
-      const res = await fetch(`api/lifts/${id}`);
-      const tempData = await res.json();
-      /*const res = await fetch(`api/week/${dayid}/${id}`); // this will need work
-      const data = await res.json();*/
-      return tempData;
-    },
-  },
 };
 </script>
